new file for clinicalInformation
